fix(reml): stop OOM right after whole-dataset load
Loading N runs into a list and np.concatenate-ing kept both the list and
the new array alive at once (~2x peak). At 100-run / 160 GB scale that
doubling OOM-killed the host right after load (kernel: total-vm 325 GB).
Preallocate one (n_voxels, n_timepoints) float32 buffer from the design's
n_timepoints and fill it run-by-run, freeing each run before the next —
peak drops to ~data + one run.

Two companion memory fixes on the same path:
- scale_to_percent_signal gains track_violations=False, which skips the
  full (n_voxels, n_timepoints) bool mask (~40 GB at whole-dataset scale)
  and derives the counts from a per-voxel accumulator instead. reml only
  reads the counts, so it opts out; scale_info is byte-identical (tested).
- load_nifti reads .nii.zst as float32 (get_fdata defaulted to float64,
  doubling per-volume RAM for no benefit).

+++ fastfuncstuff/io/afni.py
@@ -239,8 +239,11 @@ def load_nifti(filepath: str | Path) -> nib.Nifti1Image:
             img = nib.load(tmp_path)
 
             # Load data into memory and create new image to avoid lazy loading issues
-            # This ensures the temp file can be safely deleted
-            data = img.get_fdata()
+            # This ensures the temp file can be safely deleted. Read as float32:
+            # get_fdata() defaults to float64, which doubles peak RAM per volume
+            # (catastrophic for whole-dataset 4D loads) for no benefit -- callers
+            # that need float64 recast downstream.
+            data = img.get_fdata(dtype=np.float32)
             affine = img.affine
             header = img.header.copy()
